+++ newsroom/news_api/settings.py
@@ -22,8 +22,6 @@ CORE_APPS = [
     'newsroom.news_api.api_audit',
     'newsroom.news_api.news.assets.assets',
     'newsroom.upload',
-    'newsroom.news_api.news.atom.atom',
-    'newsroom.news_api.news.rss.rss',
     'newsroom.history'
 ]
